+++ src/modules/auth/clients/ClientCreateContainer.js
@@ -108,6 +108,10 @@ const ClientCreateContainer = (props) => {
 			body.client_name = "";
 		}
 
+		if (body.preferred_client_id == "" || body.preferred_client_id == undefined) {
+			delete body.preferred_client_id;
+		}
+
 		try {
 			let response = await SeaCatAuthAPI.post(`/client`, body);
 			if (response.statusText != 'OK') {
@@ -143,6 +147,7 @@ const ClientCreateContainer = (props) => {
 	const changeTemplate = (value) => {
 		resetField("client_name");
 		resetField("client_uri");
+		resetField("preferred_client_id");
 		resetField("application_type");
 		resetField("grant_types");
 		resetField("response_types");
@@ -185,6 +190,7 @@ const ClientCreateContainer = (props) => {
 										case 'redirect_uris': return(<URiInput key={key} name={key} control={control} errors={errors} append={append} remove={remove} fields={fields} labelName={t("ClientCreateContainer|Redirect URIs")}/>)
 										case 'client_name': return(<TextInput key={key} name={key} register={register} labelName={t('ClientCreateContainer|Client name')}/>)
 										case 'client_uri': return(<TextInput key={key} name={key} register={register} labelName={t('ClientCreateContainer|Client URI')}/>)
+										case 'preferred_client_id': return(<TextInput key={key} name={key} register={register} errors={errors} labelName={t('ClientCreateContainer|Preferred client ID')}/>)
 										case 'response_types': return(selectedTemplate === "Custom" && <MultiCheckbox key={key} name={key} value={value["items"]["enum"]} setValue={setValue} labelName={t('ClientCreateContainer|Response types')}/>)
 										case 'grant_types': return(selectedTemplate === "Custom" && <MultiCheckbox key={key} name={key} value={value["items"]["enum"]} setValue={setValue} labelName={t('ClientCreateContainer|Grant types')}/>)
 										case 'application_type': return(selectedTemplate === "Custom" && <SelectInput key={key} name={key} register={register} value={value["enum"]} labelName={t('ClientCreateContainer|Application type')}/>)